pyproject.toml
[project]
name = "pricing-engine"
version = "0.1.0"
description = ""
authors = [
    {name = "www",email = "pyostr@gmail.com"}
]
requires-python = "^3.13"
dependencies = [
    "asyncpg (>=0.31.0,<0.32.0)",
    "redis (>=7.2.1,<8.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "requests (>=2.32.5,<3.0.0)",
    "async-cache (>=1.2.1,<2.0.0)",
    "async-timeout (>=5.0.1,<6.0.0)",
    "cryptography (>=46.0.5,<47.0.0)",
    "sqlalchemy (>=2.0.47,<3.0.0)",
    "infisicalsdk (>=1.0.16,<2.0.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "boto3 (>=1.42.57,<2.0.0)",
    "pyjwt (>=2.11.0,<3.0.0)",
    "fastapi (>=0.133.1,<0.134.0)",
    "alembic (>=1.18.4,<2.0.0)",
    "pydantic-settings (>=2.13.1,<3.0.0)",
    "uvicorn (>=0.41.0,<0.42.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "pytest (>=9.0.2,<10.0.0)",
    "pytest-asyncio (>=1.3.0,<2.0.0)",
    "isort (>=8.0.1,<9.0.0)",
    "testcontainers[postgresql] (>=4.14.1,<5.0.0)"
]



[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
